Delete everything other than weeknotes
This is just a first stab, more significant changes are needed to turn
this from a generic memo site into a weeknotes site.  eg:

- Taxonomy should go (but year archives / feeds should stay)
- Files / URLs can be `NNN`, rather than `weeknotes-NNN`
- Build script can be significantly simplified
- Template can be re-done

The new pipeline will need adding to Concourse, the new site configuring
in Caddy.  And then when it's all set up, the original pages can be
removed from `memo.barrucadu.co.uk` (which'll also significantly speed
that up) and the old URLs redirected.